Support single-member struct serialization
We encountered this bug when trying to serialize a struct that had just
a single member. To get a failing test going, we had to add some extra
compiler flags and were then able to get a fix together. Changes to other
Serialize files were just to resolve some of these errors, mostly for
unused parameters